FC: Improve overlap checking algorithm
Instead of checking for the longest common substring, we need to store
all common substrings (length greater than 2) for checking. This is
because the longest overlap may not necessarily be the most
appropriate for merging two strings.
Due to this we now need to iterate over the list of substrings in order
to find the longest substring that either corresponds to the end and
start of the first and second compared strings, and vice versa, or is
equal to one of the compared strings. This validity check is done in its
own function